requirements.txt
streamlit==1.38.0
pandas==2.2.2
gspread==6.1.4
google-auth==2.35.0
python-dotenv==1.0.1
numpy==2.1.0
